Moved out of the __init__.py the functions: check_min_generate_level() and modify_persentages()
