+++ src/Ankimon/functions/encounter_functions.py
@@ -1,0 +1,57 @@
+from ..singletons import main_pokemon
+
+def modify_percentages(total_reviews, daily_average, player_level):
+    """
+    Modify Pokémon encounter percentages based on total reviews, player level, event modifiers, and main Pokémon level.
+    """
+    # Start with the base percentages
+    percentages = {"Baby": 2, "Legendary": 0.5, "Mythical": 0.2, "Normal": 92.3, "Ultra": 5}
+
+    # Adjust percentages based on total reviews relative to the daily average
+    review_ratio = total_reviews / daily_average if daily_average > 0 else 0
+
+    # Adjust for review progress
+    if review_ratio < 0.4:
+        percentages["Normal"] += percentages.pop("Baby", 0) + percentages.pop("Legendary", 0) + \
+                                 percentages.pop("Mythical", 0) + percentages.pop("Ultra", 0)
+    elif review_ratio < 0.6:
+        percentages["Baby"] += 2
+        percentages["Normal"] -= 2
+    elif review_ratio < 0.8:
+        percentages["Ultra"] += 3
+        percentages["Normal"] -= 3
+    else:
+        percentages["Legendary"] += 2
+        percentages["Ultra"] += 3
+        percentages["Normal"] -= 5
+
+    # Restrict access to certain tiers based on main Pokémon level
+    if main_pokemon.level:
+        # Define level thresholds for each tier
+        level_thresholds = {
+            "Ultra": 30,  # Example threshold for Ultra Pokémon
+            "Legendary": 50,  # Example threshold for Legendary Pokémon
+            "Mythical": 75  # Example threshold for Mythical Pokémon
+        }
+
+        for tier in ["Ultra", "Legendary", "Mythical"]:
+            if main_pokemon.level < level_thresholds.get(tier, float("inf")):
+                percentages[tier] = 0  # Set percentage to 0 if the level requirement isn't met
+
+    # Example modification based on player level
+    if player_level:
+        adjustment = 5  # Adjustment value for the example
+        if player_level > 10:
+            for tier in percentages:
+                if tier == "Normal":
+                    percentages[tier] = max(percentages[tier] - adjustment, 0)
+                else:
+                    percentages[tier] = percentages.get(tier, 0) + adjustment
+                    
+    # Normalize percentages to ensure they sum to 100
+    total = sum(percentages.values())
+    for tier in percentages:
+        percentages[tier] = (percentages[tier] / total) * 100 if total > 0 else 0
+    # this function gets called maybe 10 times per battle round, which is concerning. 
+    # it could be rewritten to run ONLY when the change in review ratio is detected.
+    return percentages
